+++ src/common/hacks.h
@@ -39,7 +39,8 @@
 #define ENGAGE_NO_DELAY_FOR_GARBAGE_IN_AVI  18
 #define ENGAGE_KEEP_LAST_CHAPTER_IN_MPLS    19
 #define ENGAGE_KEEP_TRACK_STATISTICS_TAGS   20
-#define ENGAGE_MAX_IDX                      20
+#define ENGAGE_ALL_I_SLICES_ARE_KEY_FRAMES  21
+#define ENGAGE_MAX_IDX                      21
 
 void engage_hacks(const std::string &hacks);
 void engage_hack(unsigned int id);