+++ src/common/mpeg4_p10.cpp
@@ -863,6 +863,7 @@ mpeg4::p10::avc_es_parser_c::avc_es_parser_c()
   , m_discard_actual_frames(false)
   , m_simple_picture_order{}
   , m_first_cleanup{true}
+  , m_all_i_slices_are_key_frames{hack_engaged(ENGAGE_ALL_I_SLICES_ARE_KEY_FRAMES)}
   , m_debug_keyframe_detection{"avc_parser|avc_keyframe_detection"}
   , m_debug_nalu_types{        "avc_parser|avc_nalu_types"}
   , m_debug_timecodes{         "avc_parser|avc_timecodes"}
@@ -1135,7 +1136,9 @@ mpeg4::p10::avc_es_parser_c::handle_slice_nalu(memory_cptr const &nalu) {
   m_incomplete_frame.m_si       =  si;
   m_incomplete_frame.m_keyframe =  m_recovery_point_valid
                                 || (   is_i_slice
-                                    && (NALU_TYPE_IDR_SLICE == si.nalu_type));
+                                    && (NALU_TYPE_IDR_SLICE == si.nalu_type))
+                                || (   is_i_slice
+                                    && m_all_i_slices_are_key_frames);
   m_incomplete_frame.m_type     =  m_incomplete_frame.m_keyframe ? 'I' : is_b_slice ? 'B' : 'P';
   m_recovery_point_valid        =  false;